Starting functions dispatch to sub-modules
