+++ vfrecovery/download_functions/__init__.py
@@ -1,0 +1,2 @@
+from armor3d import Armor3d
+from glorys import Glorys
